refactor(TP-161): complete Step 4 — update all test imports to new module locations
